ship with dep, glide, golint, and ineffassign
Tools dep, glide, golint, and ineffassign are popular among the Golang
crowd, and 'dep' is scheduled to be the default vendoring manager.
Therefore it is better to ship the image with these tools once than
them being gotten and built N times.

As Golang is needed to obtain their binaries, this became a multi-stage
Docker build. To reproduce it you'd need to build an image once using
the previous commit's Dockerfile, and then run this one.
  You need docker ≥17.05 for this.